Clean up by moving hooks into their own file.
