+++ ui/framework/src/ui-framework.ts
@@ -84,7 +84,10 @@ export * from "./ui-framework/frontstage/ModalSettingsStage";
 export * from "./ui-framework/hooks/useActiveIModelConnection";
 export * from "./ui-framework/hooks/useActiveViewport";
 export * from "./ui-framework/hooks/useAvailableUiItemsProviders";
+export * from "./ui-framework/hooks/useAnalysisAnimationDataProvider";
 export * from "./ui-framework/hooks/useFrameworkVersion";
+export * from "./ui-framework/hooks/useScheduleAnimationDataProvider";
+export * from "./ui-framework/hooks/useSolarDataProvider";
 
 export * from "./ui-framework/imodel-components/spatial-tree/SpatialContainmentTree";
 export * from "./ui-framework/imodel-components/category-tree/CategoriesTree";
